+++ src/pages/experiment/fieldExperimentUiSchema.ts
@@ -180,10 +180,11 @@ const fieldExperimentUiSchema = {
         gridCols: 2,
       },
       "ui:title": "",
-      "ui:order": ["filename", "comment_type"],
-      filename: {
-        "ui:placeholder": "e.g., permit-consultation-comments.pdf",
+      "ui:order": ["url", "comment_type", "description"],
+      url: {
+        "ui:placeholder": "URL or DOI (e.g. https://...)",
       },
+      description: textAreaWidget,
       comment_type: {
         "ui:widget": "CustomSelectWidget",
         "ui:enumNames": enumNames.PublicCommentType,